requirements.txt
python-telegram-bot==20.7
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
apscheduler==3.10.4
sqlite3